Using decimal instead of float

+++ internal/app/cash_register_app.go
@@ -36,7 +36,7 @@ func NewCashRegisterApp(
 	return a, nil
 }
 
-func (a CashRegisterApp) AddProductToCart(productCode string, quantity int) error {
+func (a CashRegisterApp) AddProductToCart(productCode string, quantity int64) error {
 	product, err := a.getProductFromCache(productCode)
 	if err != nil {
 		return err
